pyproject.toml
[tool.poetry]
name = "Mr.-Happy"
version = "0.1.2"
description = "A discord bot that interacts with server users, has movitivation commands and much more coming soon…"
authors = ["Rafael Rigo rigo.f.rafa@gmail.com"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
flask = "^1.1.2"
replit = "^3.0.0"
discord-pretty-help = "^1.3.0"
easypydb = "^0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
